spring-001
normlizacion de bd

+++ iam-core-p/iam-core/src/main/java/com/allianceever/projectERP/service/LeavesService.java
@@ -5,21 +5,19 @@ import com.allianceever.projectERP.model.dto.LeavesDto;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+
 @Service
 public interface LeavesService {
 
-    LeavesDto getByLeavesID(Integer leavesID);
-
-
+    LeavesDto getByLeavesID(Long leavesID);
 
     LeavesDto create(LeavesDto leavesDto);
 
-    LeavesDto update(Integer leavesID,LeavesDto leavesDto);
-
+    LeavesDto update(Long leavesID, LeavesDto leavesDto);
 
     List<LeavesDto> getAllLeavesOrderedByDate();
 
     List<LeavesDto> getAllLeavesByUsernameOrderedByDate(String username);
 
-    void delete(Integer leavesID);
+    void delete(Long leavesID);
 }
